refactor(compact): extract source test modules
- Move compact unit tests into sibling *_test.rs files
- Keep compaction runtime modules separate from test bodies
